[IMP] Update module structure and add Travis * Move files from Reverend subfolder, which was historically part of larger project * Change COPYING to License.md and update with markdown LGPL * Add travis file (lint disabled for now) * Update readme with RST * Add gitignore, coveragerc, codeclimate * Use standardized PQT setup file
