+++ .codeclimate.yml
@@ -1,0 +1,2 @@
+languages:
+  Python: true
