+++ .coveragerc
@@ -1,0 +1,19 @@
+[run]
+
+branch = True
+
+[report]
+
+exclude_lines =
+    pragma: no cover
+    def __repr__
+
+ignore_errors = True
+
+include =
+    */reverend/*
+
+omit =
+    */tests/*
+    setup.py
+    tests.py
