--- Reverend/MANIFEST.in
@@ -1,6 +1,0 @@
-include README.txt
-include LICENSE
-include changelog.txt
-
-recursive-include examples *.txt
-recursive-include examples *.py
