--- Reverend/NEWS.txt
@@ -1,3 +1,0 @@
-0.4.0 (2009-11-25):
-  - Fix bugs in `EmailClassifier` which resulted in exceptions from `train`
-    and `guess`.
