+++ reverend/ui/trainer.py
